refactor(lsp): break module_graph.go into smaller focused files
- Split 1613-line module_graph.go into 6 logical files:
  - module_graph_interfaces.go: Interface definitions and constants
  - module_graph_lazy.go: Lazy loading functionality
  - module_graph_metrics.go: Metrics collection
  - module_graph_parsers.go: Export parsing logic
  - module_graph_tracking.go: Export and dependency tracking
- Maintain package cohesion while improving maintainability
- Update documentation with tree-sitter concurrency insights
- Remove unused cachedQueryMatchers variable
