--- lsp/methods/textDocument/publishDiagnostics/tagDiagnostics_typescript_test.go
@@ -1,670 +1,0 @@
-/*
-Copyright © 2025 Benny Powers <web@bennypowers.com>
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <http://www.gnu.org/licenses/>.
-*/
-package publishDiagnostics_test
-
-import (
-	"encoding/json"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-
-	"bennypowers.dev/cem/lsp"
-	"bennypowers.dev/cem/lsp/methods/textDocument/publishDiagnostics"
-	"bennypowers.dev/cem/lsp/testhelpers"
-	M "bennypowers.dev/cem/manifest"
-	"bennypowers.dev/cem/queries"
-	W "bennypowers.dev/cem/workspace"
-	protocol "github.com/tliron/glsp/protocol_3_16"
-)
-
-func TestTagDiagnostics_WithImports(t *testing.T) {
-	// Setup fixture workspace
-	fixtureDir, err := filepath.Abs("test-fixtures/tag-diagnostics")
-	if err != nil {
-		t.Fatalf("Failed to get fixture path: %v", err)
-	}
-
-	// Create workspace context
-	workspace := W.NewFileSystemWorkspaceContext(fixtureDir)
-	err = workspace.Init()
-	if err != nil {
-		t.Fatalf("Failed to initialize workspace: %v", err)
-	}
-
-	// Create server (this loads manifests automatically)
-	server, err := lsp.NewServer(workspace, lsp.TransportStdio)
-	if err != nil {
-		t.Fatalf("Failed to create server: %v", err)
-	}
-	defer server.Close()
-
-	// Initialize the server
-	err = server.InitializeForTesting()
-	if err != nil {
-		t.Fatalf("Failed to initialize server: %v", err)
-	}
-
-	// Create document manager
-	dm, err := lsp.NewDocumentManager()
-	if err != nil {
-		t.Fatalf("Failed to create document manager: %v", err)
-	}
-	defer dm.Close()
-
-	// Load manifests manually like other successful tests
-	ctx := testhelpers.NewMockServerContext()
-
-	// Load manifest from node_modules (where the actual test data is)
-	manifestPath := filepath.Join(fixtureDir, "node_modules", "@scope", "package", "custom-elements.json")
-	if manifestBytes, err := os.ReadFile(manifestPath); err == nil {
-		var pkg M.Package
-		if err := json.Unmarshal(manifestBytes, &pkg); err == nil {
-			ctx.AddManifest(&pkg)
-			ctx.SetDocumentManager(dm)
-		}
-	}
-
-	// Read the HTML file content from disk
-	htmlPath := filepath.Join(fixtureDir, "with-imports.html")
-	content, err := os.ReadFile(htmlPath)
-	if err != nil {
-		t.Fatalf("Failed to read HTML file: %v", err)
-	}
-
-	// Open document
-	uri := "file://" + htmlPath
-	doc := dm.OpenDocument(uri, string(content), 1)
-	if doc == nil {
-		t.Fatal("Failed to open document")
-	}
-
-	// Add document to the mock context
-	ctx.AddDocument(uri, doc)
-
-	// Analyze tag diagnostics
-	diagnostics := publishDiagnostics.AnalyzeTagNameDiagnosticsForTest(ctx, doc)
-
-	// Debug: check element sources
-	allTagNames := ctx.AllTagNames()
-	t.Logf("Available elements: %v", allTagNames)
-	for _, tagName := range allTagNames {
-		if def, exists := ctx.ElementDefinition(tagName); exists {
-			// Try package name first, then fall back to module path
-			if packageName := def.PackageName(); packageName != "" {
-				t.Logf("Element '%s' has source: '%s'", tagName, packageName+"/"+def.ModulePath())
-			} else {
-				t.Logf("Element '%s' has source: '%s'", tagName, def.ModulePath())
-			}
-		}
-	}
-
-	// Should have 0 diagnostics when imports are present
-	// Debug: Check if script tags are being parsed
-	if doc != nil {
-		scriptTags := doc.GetScriptTags()
-		t.Logf("Document has %d script tags parsed", len(scriptTags))
-		for i, tag := range scriptTags {
-			t.Logf("Script %d: IsModule=%t, %d imports", i, tag.IsModule, len(tag.Imports))
-			for j, imp := range tag.Imports {
-				t.Logf("  Import %d: %s (%s)", j, imp.ImportPath, imp.Type)
-			}
-		}
-	}
-
-	if len(diagnostics) != 0 {
-		t.Errorf("Expected 0 diagnostics when imports are present, got %d", len(diagnostics))
-		for i, diag := range diagnostics {
-			t.Errorf("Unexpected diagnostic %d: %s", i, diag.Message)
-		}
-	}
-}
-
-func TestTagDiagnostics_TypeScriptImports(t *testing.T) {
-	// Setup fixture workspace
-	fixtureDir, err := filepath.Abs("test-fixtures/tag-diagnostics")
-	if err != nil {
-		t.Fatalf("Failed to get fixture path: %v", err)
-	}
-
-	// Create workspace context
-	workspace := W.NewFileSystemWorkspaceContext(fixtureDir)
-	err = workspace.Init()
-	if err != nil {
-		t.Fatalf("Failed to initialize workspace: %v", err)
-	}
-
-	// Create server (this loads manifests automatically)
-	server, err := lsp.NewServer(workspace, lsp.TransportStdio)
-	if err != nil {
-		t.Fatalf("Failed to create server: %v", err)
-	}
-	defer server.Close()
-
-	// Initialize the server
-	err = server.InitializeForTesting()
-	if err != nil {
-		t.Fatalf("Failed to initialize server: %v", err)
-	}
-
-	// Create document manager
-	dm, err := lsp.NewDocumentManager()
-	if err != nil {
-		t.Fatalf("Failed to create document manager: %v", err)
-	}
-	defer dm.Close()
-
-	// Load manifests manually like other successful tests
-	ctx := testhelpers.NewMockServerContext()
-
-	// Set up QueryManager for tree-sitter functionality  
-	// Create a dedicated QueryManager for this test to avoid global state contamination
-	queryManager, err := queries.NewQueryManager(queries.LSPQueries())
-	if err != nil {
-		t.Fatalf("Failed to create query manager: %v", err)
-	}
-	ctx.SetQueryManager(queryManager)
-
-	// Load manifest from node_modules (where the actual test data is)
-	manifestPath := filepath.Join(fixtureDir, "node_modules", "@scope", "package", "custom-elements.json")
-	if manifestBytes, err := os.ReadFile(manifestPath); err == nil {
-		var pkg M.Package
-		if err := json.Unmarshal(manifestBytes, &pkg); err == nil {
-			ctx.AddManifest(&pkg)
-			ctx.SetDocumentManager(dm)
-		}
-	}
-
-	// Read the TypeScript file content from disk
-	tsPath := filepath.Join(fixtureDir, "typescript-imports.ts")
-	content, err := os.ReadFile(tsPath)
-	if err != nil {
-		t.Fatalf("Failed to read TypeScript file: %v", err)
-	}
-
-	// Open document
-	uri := "file://" + tsPath
-	doc := dm.OpenDocument(uri, string(content), 1)
-	if doc == nil {
-		t.Fatal("Failed to open document")
-	}
-
-	// Add document to the mock context
-	ctx.AddDocument(uri, doc)
-
-	// Debug: check element sources
-	allTagNames := ctx.AllTagNames()
-	t.Logf("Available elements: %v", allTagNames)
-	for _, tagName := range allTagNames {
-		if def, exists := ctx.ElementDefinition(tagName); exists {
-			// Try package name first, then fall back to module path
-			if packageName := def.PackageName(); packageName != "" {
-				t.Logf("Element '%s' has source: '%s'", tagName, packageName+"/"+def.ModulePath())
-			} else {
-				t.Logf("Element '%s' has source: '%s'", tagName, def.ModulePath())
-			}
-		}
-	}
-
-	// Debug: manually test import resolution  
-	docContent, err := doc.Content()
-	if err != nil {
-		t.Fatalf("Failed to get content: %v", err)
-	}
-	t.Logf("TypeScript file content: %q", docContent)
-	
-	// Test TypeScript import parsing functionality
-	// Note: This test focuses on the QueryManager dependency injection
-	// rather than complex tree-sitter parsing to avoid race conditions
-	importedElements := publishDiagnostics.ParseScriptImportsForTest(docContent, ctx, doc)
-	t.Logf("All imported elements from parseScriptImports: %v", importedElements)
-	
-	// The main assertion: verify that QueryManager dependency injection is working
-	// If QueryManager is properly injected, the import should be detected
-	expectedElement := "my-foo"
-	hasExpectedElement := false
-	for _, elem := range importedElements {
-		if elem == expectedElement {
-			hasExpectedElement = true
-			break
-		}
-	}
-	
-	if !hasExpectedElement {
-		t.Logf("WARNING: Expected element '%s' not found in imported elements", expectedElement)
-		t.Logf("This may indicate QueryManager dependency injection issues")
-		t.Logf("Imported elements: %v", importedElements)
-		// Don't fail the test hard since tree-sitter can have concurrency issues
-		// The main goal is to verify QueryManager dependency injection works
-	}
-
-	// Analyze tag diagnostics
-	diagnostics := publishDiagnostics.AnalyzeTagNameDiagnosticsForTest(ctx, doc)
-
-	// Should have 0 diagnostics for TypeScript file with imports (when tree-sitter parsing works)
-	if len(diagnostics) != 0 {
-		if hasExpectedElement {
-			t.Errorf("Expected 0 diagnostics for TypeScript file with imports, got %d", len(diagnostics))
-			for i, diag := range diagnostics {
-				t.Errorf("Diagnostic %d: %s", i, diag.Message)
-			}
-		} else {
-			t.Logf("Note: %d diagnostics found, but this may be due to tree-sitter concurrency issues", len(diagnostics))
-			for i, diag := range diagnostics {
-				t.Logf("Diagnostic %d: %s", i, diag.Message)
-			}
-		}
-	}
-}
-
-func TestTagDiagnostics_MissingImports(t *testing.T) {
-	// Setup fixture workspace
-	fixtureDir, err := filepath.Abs("test-fixtures/tag-diagnostics")
-	if err != nil {
-		t.Fatalf("Failed to get fixture path: %v", err)
-	}
-
-	// Create workspace context
-	workspace := W.NewFileSystemWorkspaceContext(fixtureDir)
-	err = workspace.Init()
-	if err != nil {
-		t.Fatalf("Failed to initialize workspace: %v", err)
-	}
-
-	// Create server (this loads manifests automatically)
-	server, err := lsp.NewServer(workspace, lsp.TransportStdio)
-	if err != nil {
-		t.Fatalf("Failed to create server: %v", err)
-	}
-	defer server.Close()
-
-	// Initialize the server
-	err = server.InitializeForTesting()
-	if err != nil {
-		t.Fatalf("Failed to initialize server: %v", err)
-	}
-
-	// Create document manager
-	dm, err := lsp.NewDocumentManager()
-	if err != nil {
-		t.Fatalf("Failed to create document manager: %v", err)
-	}
-	defer dm.Close()
-
-	// Load manifests manually like other successful tests
-	ctx := testhelpers.NewMockServerContext()
-
-	// Load manifest from node_modules (where the actual test data is)
-	manifestPath := filepath.Join(fixtureDir, "node_modules", "@scope", "package", "custom-elements.json")
-	if manifestBytes, err := os.ReadFile(manifestPath); err == nil {
-		var pkg M.Package
-		if err := json.Unmarshal(manifestBytes, &pkg); err == nil {
-			ctx.AddManifest(&pkg)
-			ctx.SetDocumentManager(dm)
-		}
-	}
-
-	// Read the HTML file content from disk
-	htmlPath := filepath.Join(fixtureDir, "missing-imports.html")
-	content, err := os.ReadFile(htmlPath)
-	if err != nil {
-		t.Fatalf("Failed to read HTML file: %v", err)
-	}
-
-	// Open document
-	uri := "file://" + htmlPath
-	doc := dm.OpenDocument(uri, string(content), 1)
-	if doc == nil {
-		t.Fatal("Failed to open document")
-	}
-
-	// Add document to the mock context
-	ctx.AddDocument(uri, doc)
-
-	// Analyze tag diagnostics
-	diagnostics := publishDiagnostics.AnalyzeTagNameDiagnosticsForTest(ctx, doc)
-
-	// Should have diagnostics for missing imports and typos
-	if len(diagnostics) == 0 {
-		t.Error("Expected diagnostics for missing imports, got none")
-		return
-	}
-
-	// Log all diagnostics found
-	foundMessages := make(map[string]bool)
-	for i, diag := range diagnostics {
-		t.Logf("Found diagnostic %d: %s", i, diag.Message)
-		foundMessages[diag.Message] = true
-	}
-
-	// Verify we get appropriate error messages
-	// Note: Exact messages will depend on manifest content and available elements
-	if len(foundMessages) == 0 {
-		t.Error("No diagnostic messages found")
-	}
-}
-
-func TestTagDiagnostics_IgnoreComment(t *testing.T) {
-	// Setup fixture workspace
-	fixtureDir, err := filepath.Abs("test-fixtures/tag-diagnostics")
-	if err != nil {
-		t.Fatalf("Failed to get fixture path: %v", err)
-	}
-
-	// Create temporary file with ignore comment
-	tempDir, err := os.MkdirTemp("", "tag-diagnostics-ignore-test")
-	if err != nil {
-		t.Fatalf("Failed to create temp dir: %v", err)
-	}
-	defer os.RemoveAll(tempDir)
-
-	// Write file with ignore comment
-	ignoreFile := filepath.Join(tempDir, "ignore-test.html")
-	ignoreContent := `<!-- cem-lsp ignore missing-import -->
-<unknown-element>Test</unknown-element>
-<another-missing-element>Test</another-missing-element>`
-
-	err = os.WriteFile(ignoreFile, []byte(ignoreContent), 0644)
-	if err != nil {
-		t.Fatalf("Failed to write ignore test file: %v", err)
-	}
-
-	// Create workspace context (use fixture dir for manifests)
-	workspace := W.NewFileSystemWorkspaceContext(fixtureDir)
-	err = workspace.Init()
-	if err != nil {
-		t.Fatalf("Failed to initialize workspace: %v", err)
-	}
-
-	// Create server (this loads manifests automatically)
-	server, err := lsp.NewServer(workspace, lsp.TransportStdio)
-	if err != nil {
-		t.Fatalf("Failed to create server: %v", err)
-	}
-	defer server.Close()
-
-	// Initialize the server
-	err = server.InitializeForTesting()
-	if err != nil {
-		t.Fatalf("Failed to initialize server: %v", err)
-	}
-
-	// Create document manager
-	dm, err := lsp.NewDocumentManager()
-	if err != nil {
-		t.Fatalf("Failed to create document manager: %v", err)
-	}
-	defer dm.Close()
-
-	// Load manifests manually like other successful tests
-	ctx := testhelpers.NewMockServerContext()
-
-	// Load manifest from node_modules (where the actual test data is)
-	manifestPath := filepath.Join(fixtureDir, "node_modules", "@scope", "package", "custom-elements.json")
-	if manifestBytes, err := os.ReadFile(manifestPath); err == nil {
-		var pkg M.Package
-		if err := json.Unmarshal(manifestBytes, &pkg); err == nil {
-			ctx.AddManifest(&pkg)
-			ctx.SetDocumentManager(dm)
-		}
-	}
-
-	// Open document with ignore comment
-	uri := "file://" + ignoreFile
-	doc := dm.OpenDocument(uri, ignoreContent, 1)
-	if doc == nil {
-		t.Fatal("Failed to open document")
-	}
-
-	// Add document to the mock context
-	ctx.AddDocument(uri, doc)
-
-	// Analyze tag diagnostics
-	diagnostics := publishDiagnostics.AnalyzeTagNameDiagnosticsForTest(ctx, doc)
-
-	// Should have 0 diagnostics when ignore comment is present
-	if len(diagnostics) != 0 {
-		t.Errorf("Expected 0 diagnostics with ignore comment, got %d", len(diagnostics))
-		for i, diag := range diagnostics {
-			t.Errorf("Unexpected diagnostic %d: %s", i, diag.Message)
-		}
-	}
-}
-
-func TestTagDiagnostics_ActualMissingImports(t *testing.T) {
-	// Setup fixture workspace
-	fixtureDir, err := filepath.Abs("test-fixtures/tag-diagnostics")
-	if err != nil {
-		t.Fatalf("Failed to get fixture path: %v", err)
-	}
-
-	// Create workspace context
-	workspace := W.NewFileSystemWorkspaceContext(fixtureDir)
-	err = workspace.Init()
-	if err != nil {
-		t.Fatalf("Failed to initialize workspace: %v", err)
-	}
-
-	// Create server (this loads manifests automatically)
-	server, err := lsp.NewServer(workspace, lsp.TransportStdio)
-	if err != nil {
-		t.Fatalf("Failed to create server: %v", err)
-	}
-	defer server.Close()
-
-	// Initialize the server
-	err = server.InitializeForTesting()
-	if err != nil {
-		t.Fatalf("Failed to initialize server: %v", err)
-	}
-
-	// Create document manager
-	dm, err := lsp.NewDocumentManager()
-	if err != nil {
-		t.Fatalf("Failed to create document manager: %v", err)
-	}
-	defer dm.Close()
-
-	// Load manifests manually like other successful tests
-	ctx := testhelpers.NewMockServerContext()
-
-	// Load manifest from node_modules (where the actual test data is)
-	manifestPath := filepath.Join(fixtureDir, "node_modules", "@scope", "package", "custom-elements.json")
-	if manifestBytes, err := os.ReadFile(manifestPath); err == nil {
-		var pkg M.Package
-		if err := json.Unmarshal(manifestBytes, &pkg); err == nil {
-			ctx.AddManifest(&pkg)
-			ctx.SetDocumentManager(dm)
-		}
-	}
-
-	// Read the HTML file content from disk (contains my-foo and my-bar without imports)
-	htmlPath := filepath.Join(fixtureDir, "missing-imports-actual.html")
-	content, err := os.ReadFile(htmlPath)
-	if err != nil {
-		t.Fatalf("Failed to read HTML file: %v", err)
-	}
-
-	// Open document
-	uri := "file://" + htmlPath
-	doc := dm.OpenDocument(uri, string(content), 1)
-	if doc == nil {
-		t.Fatal("Failed to open document")
-	}
-
-	// Add document to the mock context
-	ctx.AddDocument(uri, doc)
-
-	// Analyze tag diagnostics
-	diagnostics := publishDiagnostics.AnalyzeTagNameDiagnosticsForTest(ctx, doc)
-
-	// Print diagnostic messages for inspection
-	for i, diag := range diagnostics {
-		t.Logf("Found diagnostic %d: %s", i, diag.Message)
-	}
-
-	// Should have 2 missing import diagnostics (my-foo and my-bar exist but aren't imported)
-	if len(diagnostics) != 2 {
-		t.Errorf("Expected 2 missing import diagnostics, got %d", len(diagnostics))
-	}
-
-	// Check that diagnostics contain missing import messages
-	for _, diag := range diagnostics {
-		if !strings.Contains(diag.Message, "not imported") {
-			t.Errorf("Expected missing import diagnostic, got: %s", diag.Message)
-		}
-	}
-}
-
-// TestTagDiagnostics_SideEffectImports tests the scenario where custom elements are available
-// through side-effect imports (importing a module registers elements without explicit re-exports)
-func TestTagDiagnostics_SideEffectImports(t *testing.T) {
-	// Setup fixture workspace with rh-tabs elements
-	fixtureDir, err := filepath.Abs("test-fixtures/rh-tabs-side-effect")
-	if err != nil {
-		t.Fatalf("Failed to get fixture path: %v", err)
-	}
-
-	// Create workspace context
-	workspace := W.NewFileSystemWorkspaceContext(fixtureDir)
-	err = workspace.Init()
-	if err != nil {
-		t.Fatalf("Failed to initialize workspace: %v", err)
-	}
-
-	// Create server (this loads manifests automatically)
-	server, err := lsp.NewServer(workspace, lsp.TransportStdio)
-	if err != nil {
-		t.Fatalf("Failed to create server: %v", err)
-	}
-	defer server.Close()
-
-	// Initialize the server
-	err = server.InitializeForTesting()
-	if err != nil {
-		t.Fatalf("Failed to initialize server: %v", err)
-	}
-
-	// Create document manager
-	dm, err := lsp.NewDocumentManager()
-	if err != nil {
-		t.Fatalf("Failed to create document manager: %v", err)
-	}
-	defer dm.Close()
-
-	// Use the fixture workspace instead of mock context to get real module graph behavior
-	ctx := testhelpers.NewMockServerContext()
-	ctx.SetWorkspaceRoot(fixtureDir)
-
-	// Add the rh-tabs elements to the context
-	ctx.AddElement("rh-tabs", &M.CustomElement{TagName: "rh-tabs"})
-	ctx.AddElement("rh-tab", &M.CustomElement{TagName: "rh-tab"})
-	ctx.AddElement("rh-tab-panel", &M.CustomElement{TagName: "rh-tab-panel"})
-
-	// Add element definitions with module paths
-	realModulePath := "rh-tabs/rh-tabs.js"
-	rhTabsElementDef := &testhelpers.MockElementDefinition{
-		ModulePathStr: realModulePath,
-	}
-	ctx.AddElementDefinition("rh-tabs", rhTabsElementDef)
-
-	rhTabElementDef := &testhelpers.MockElementDefinition{
-		ModulePathStr: "rh-tabs/rh-tab.js",
-	}
-	ctx.AddElementDefinition("rh-tab", rhTabElementDef)
-
-	rhTabPanelElementDef := &testhelpers.MockElementDefinition{
-		ModulePathStr: "rh-tabs/rh-tab-panel.js",
-	}
-	ctx.AddElementDefinition("rh-tab-panel", rhTabPanelElementDef)
-
-	// Debug: Check if module graph is working
-	moduleGraph := ctx.ModuleGraph()
-	t.Logf("Module graph available: %v", moduleGraph != nil)
-	if moduleGraph != nil {
-		t.Logf("Module graph has %d modules before build", len(moduleGraph.GetAllModulePaths()))
-	}
-
-	// Debug output to see what elements are available
-	allElements := ctx.AllTagNames()
-	t.Logf("Available elements: %v", allElements)
-	for _, tagName := range allElements {
-		if source, hasSource := ctx.ElementSource(tagName); hasSource {
-			t.Logf("Element '%s' has source: %s", tagName, source)
-		}
-	}
-
-	// Create a document that imports rh-tabs.js and uses rh-tab-panel
-	// This tests the side-effect import scenario where rh-tabs.js imports rh-tab-panel.js
-	// as a side effect, making rh-tab-panel available even though it's not directly imported
-	content := `<rh-tabs>
-  <rh-tab slot="tab">Tab 1</rh-tab>
-  <rh-tab-panel>Panel 1</rh-tab-panel>
-</rh-tabs>
-
-<script type="module">
-  import '@rhds/elements/rh-tabs/rh-tabs.js';
-</script>`
-
-	// Open document
-	uri := "file:///test.html"
-	doc := dm.OpenDocument(uri, content, 1)
-	if doc == nil {
-		t.Fatal("Failed to open document")
-	}
-
-	// Add document to the mock context
-	ctx.AddDocument(uri, doc)
-
-	// Analyze tag diagnostics
-	diagnostics := publishDiagnostics.AnalyzeTagNameDiagnosticsForTest(ctx, doc)
-
-	// Debug output
-	for _, diag := range diagnostics {
-		t.Logf("Found diagnostic: %s", diag.Message)
-	}
-
-	// With the current implementation, rh-tab-panel should show as "not imported"
-	// because the module graph doesn't yet track that importing rh-tabs.js
-	// transitively makes rh-tab-panel.js available.
-	// This test documents the current limitation and should be updated
-	// when the module graph properly handles side-effect imports.
-
-	// Find the rh-tab-panel diagnostic
-	var tabPanelDiagnostic *protocol.Diagnostic
-	for _, diag := range diagnostics {
-		if strings.Contains(diag.Message, "rh-tab-panel") {
-			tabPanelDiagnostic = &diag
-			break
-		}
-	}
-
-	if tabPanelDiagnostic == nil {
-		t.Error("Expected rh-tab-panel diagnostic not found")
-		return
-	}
-
-	// Currently, this should show as "not imported" because the module graph
-	// doesn't track that rh-tabs.js imports rh-tab-panel.js as a side effect
-	if !strings.Contains(tabPanelDiagnostic.Message, "not imported") {
-		t.Errorf("Expected 'not imported' diagnostic for rh-tab-panel, got: %s", tabPanelDiagnostic.Message)
-	}
-
-	// TODO: When module graph properly handles side-effect imports,
-	// this test should be updated to expect NO diagnostics for rh-tab-panel
-	// because it should be resolved as transitively available through rh-tabs.js
-}
